Move the samples of the sliceRows operations to the `samples` module
