+++ docs/StardustDocs/resources/api/pivot/pivotInward_properties.html
@@ -459,7 +459,7 @@ summary {
 /*<!--*/
 call_DataFrame(function() { DataFrame.addTable({ cols: [{ name: "<span title=\"true: api.FormattedFrame<Any?>\" style=\"color:#f44336;font-family:ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace\">true</span>", children: [], rightAlign: false, values: [{ frameId: 1, value: "<b>DataFrame 5 x 5</b>" }] }, 
 { name: "<span title=\"false: api.FormattedFrame<Any?>\" style=\"color:#f44336;font-family:ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace\">false</span>", children: [], rightAlign: false, values: [{ frameId: 2, value: "<b>DataFrame 5 x 5</b>" }] }, 
-{ name: "<span title=\"isHappy: DataRow<*>\" style=\"color:#f44336;font-family:ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace\">isHappy</span>", children: [0, 1], rightAlign: false, values: ["<span class=\"formatted\" title=\"true: org.jetbrains.kotlinx.dataframe.api.FormattedFrame@27da994b\nfalse: org.jetbrains.kotlinx.dataframe.api.FormattedFrame@24a4e2c5\"><span class=\"structural\">{ </span><span class=\"structural\">true: </span>org.jetbrains.kotlinx<span class=\"structural\">...</span><span class=\"structural\">, </span><span class=\"structural\">f...</span><span class=\"structural\"> }</span></span>"] },
+{ name: "<span title=\"isHappy: DataRow<*>\" style=\"color:#f44336;font-family:ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', 'Courier New', monospace\">isHappy</span>", children: [0, 1], rightAlign: false, values: ["<span class=\"formatted\" title=\"true: org.jetbrains.kotlinx.dataframe.api.FormattedFrame@21390938\nfalse: org.jetbrains.kotlinx.dataframe.api.FormattedFrame@67e0fd6d\"><span class=\"structural\">{ </span><span class=\"structural\">true: </span>org.jetbrains.kotlinx<span class=\"structural\">...</span><span class=\"structural\">, </span><span class=\"structural\">f...</span><span class=\"structural\"> }</span></span>"] }, 
 ], id: 0, rootId: 0, totalRows: 1 } ) });
 /*-->*/
