+++ docs/StardustDocs/topics/_shadow_resources.md
@@ -235,23 +235,15 @@
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.distinct.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.distinctBy.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.distinctColumns.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.drop.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropLast.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNA.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNaNs.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropNulls.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropWhere.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.dropWhile.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.filter.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.getColumnsByName.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.getRowByCondition.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.getRowByIndex.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.getSeveralRowsByIndices.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.getSeveralRowsByRanges.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.select.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.take.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.takeLast.html"></resource>
-<resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.takeWhile.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Access.xs.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Analyze.columnsFor.html"></resource>
 <resource src="org.jetbrains.kotlinx.dataframe.samples.api.Analyze.countAggregation.html"></resource>
@@ -356,3 +348,13 @@
 <resource src="countGroupBy_properties.html"></resource>
 <resource src="countPivot_properties.html"></resource>
 <resource src="countPivotGroupBy_properties.html"></resource>
+<resource src="getSeveralRowsByIndices.html"></resource>
+<resource src="getSeveralRowsByRanges1.html"></resource>
+<resource src="getSeveralRowsByRanges2.html"></resource>
+<resource src="take.html"></resource>
+<resource src="takeLast.html"></resource>
+<resource src="takeWhile.html"></resource>
+<resource src="drop.html"></resource>
+<resource src="dropLast1.html"></resource>
+<resource src="dropLast2.html"></resource>
+<resource src="dropWhile.html"></resource>
